feat: add waymo datasets on Mega, subject to change

+++ mega_nerf/rendering.py
@@ -40,7 +40,7 @@ def render_rays(nerf: nn.Module,
         rays_d = rays_d.view(rays_d.shape[0], 1, rays_d.shape[1])
 
         if rays_with_bg.shape[0] > 0:
-            last_delta[rays_with_bg, 0] = fg_far[rays_with_bg]
+            last_delta[rays_with_bg, 0] = fg_far[rays_with_bg].float()
 
             far = torch.minimum(far.squeeze(), fg_far).unsqueeze(-1)
